Clean up streaming adapter
